marina-rest-api/requirements.txt
Django==1.11.2
gunicorn==19.7.1
djangorestframework
